time: adjust timer and timeout utils and calculations

+++ src/address.c
@@ -1684,7 +1684,7 @@ ni_lifetime_left(unsigned int lifetime, const struct timeval *acquired, const st
 
 	if (timercmp(current, acquired, >)) {
 		timersub(current, acquired, &dif);
-		if (lifetime >= dif.tv_sec) {
+		if ((unsigned long)lifetime >= (unsigned long)dif.tv_sec) {
 			lifetime -= dif.tv_sec;
 			return lifetime;
 		}
